logging: allow subclasses to override json serialization.
It can be very useful to consumers of the dockerflow library to override
the json serialization used by the JsonLogFormatter. This allows custom
serialization of specific object types (Currently handled by the
SafeJSONEncoder).

My specific use case for this functionality is allowing dockerflow
library users to create a "Pretty" formatter to be used when in DEBUG
mode, such as the following:

```
class PrettyJsonLogFormatter(JsonLogFormatter):
    """Log formatter that outputs human-readable json."""

    def serialize(obj):
        """Serialize to sorted multi-line json."""
        return json.dumps(obj, cls=SafeJSONEncoder, sort_keys=True, indent=2)
```