Update Django docs (#109)
* Specify that the middleware serves the views

Since the DockerflowMiddleware serves the health monitoring views, it is
unneccessary to add them to urls.py

* Expand DOCKERFLOW_VERSION_CALLBACK docs

Add links to other documentation. Fix the signature of the callback
function. Move so that the section is alpha-sorted.

Fixup version callback

* Add DOCKERFLOW_REQUEST_ID_HEADER_NAME docs

* Add DOCKERFLOW_SUMMARY_LOG_QUERYSTRING docs

* Add DEBUG docs

* Add SILENCED_SYSTEM_CHECKS docs

* Add DOCKERFLOW_HEARTBEAT_FAILED_STATUS_CODE docs

* MIDDLEWARE_CLASSES was removed in Django 2.0, 2017

* Add dockerflow.checks.register to log config

* Use dockerflow as logger name

These logs should propagate up, so use more general logging level.

* Fix name of RequestIdLogFilter

* Add SECURE_REDIRECT_EXEMPT docs

* More MIDDLEWARE_CLASSES removal

* Less bear 🐻

* Fix name of RequestIdLogFilter